=== WooCommerce Ajax Cart Plugin ===
WooCommerce AJAX Cart plugin is a WordPress Plugin that changes the default behavior of WooCommerte Cart Page, allowing a buyer to see the Total price calculation when change the Quantity of a product, without need to manually click on "Update cart" button.
This improves the user experience and commodity when purchasing a product. No other hacks/code/theme changes is needed, this functionality is added when the plugin is activated.